backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
python-jose
pydantic
psycopg2-binary
email-validator
python-multipart
bcrypt
